Refactor the langpack symlink creation stuff at startup
We try to create a symlink for the short locale first if we can,
and then try the longer locale.  The code to do this is extremely
similar, so create a bash function to do the work for each.

This will make it easier to switch to a packed langpack extension
format in the future, which will help with startup performance.