Use upstream's version of our patch to set the default browser
When setting the default browser, we want to set it to 'firefox %s'
instead of '/usr/lib{,64}/firefox-4/firefox %s'

firefox-default.patch did the job for us for a bit, but upstream now
has a patch to do this which we can now use.  The patch will check
to see if MOZ_APP_LAUNCHER is passed, and if so, it will use only
the basename when setting the default browser.  We already set
MOZ_APP_LAUNCHER, so just need to switch to using the patch.

From https://bugzilla.mozilla.org/show_bug.cgi?id=611953
Patch 1 - Use MOZ_APP_LAUNCHER for default browser executable (v3, un-bitrotted)