Updates for GNOME 3
Taken from https://bugzilla.mozilla.org/show_bug.cgi?id=611953
Patches	2-5.

Not all	patches	have positive review yet; there's a question of
whether to ensure a gsettings key exists, due to picture-filename
changing to picture-uri in the middle of the review process.

This is	something we can control in the	distribution, though, and
don't need to be dynamic about it.