Update content and referenced paths in some broken patch files.
Zbigniew (aka Zbyszek) was a lot of help, once again, this time
interpreting the error messages rpmbuild was giving me. The actual
content for some of the patch files is out-of-date, out of sync with the
source code it patches; and for some of the patch files, the paths in
the headers are messed up.

Patch files *paths* could be easily fixed and more easily managed, if we
were more consistent about where the paths are supposed to start. I'll
be updating the path headers in the rest of the patch files next.

(I'm also occasionally updating the firefox version and commit ref. But
since this doesn't even build, currently, I'm not too worried about
always building the latest source code.)