Normalize paths in patch file headers. Skip patches with outdated content.
All paths referenced in the patch files should now be relative to the
%{tarballdir} directory (defined in the .spec file), since that's the
current working directory when applying patches.

The filename suffixes used when actually applying the patches were, I
think, intended to describe the applied patch. But they don't really
tell you much about what's happening in the patch; a tilde is a much
more common backup filename suffix, and easier to catch when glancing
over a directory.

patch215 and patch500 are so out of sync with the files they're supposed
to be patching, that I'm not sure if they're even relevant anymore. I'm
commenting them out for now, but may remove them later.