Simplify config.status shebang line, which ultimately calls Python2.
In the %{tarballdir}, the build process generates a file called
`objdir/config.status`, which has a shebang line calling
`objdir/_virtualenv/bin/python` which calls python2.7. The shebang line
in config.status exceeds the character limit set in the kernel (see `man
execve`), and it's an awfully roundabout way of evoking python. So I'm
replacing it with call to `env python` instead.