Trying to make macros/variables more flexible.
Looks like I needed a macro for the dateless version number
(version_short) after all. Variables were renamed in GET_SOURCE.bash to
try to better align with the corresponding variables in the spec file.

I think the Firefox source and/or compiler must have some paths or
filenames hard-coded, because a `sed` line here in the spec, and the
%files section, expect files to be installed to BUILDROOT/.../lib64
/firefox-dev/, but they're getting installed to .../firefox/ instead. I
replaced some instances of "firefox" with "%{name}" to help with this,
but the problem is mostly in the lib directories, and I don't see any
way of changing those paths from the spec file. I may need to patch the
paths later or something.