Correct names and paths for the %files check.
The RPM build has been failing just before it could finish, because
packaged files were seemingly missing.

A big part of the problem I think lies in the fact that the build
scripts by Mozilla take for granted that the packaged application is
going to be named Firefox (unless you use the compile options flag to
disable official branding). So many of the library files were still
getting placed in a lib directory simply named "firefox" instead of
firefox-dev or firefox-aurora or whatever. Simply moving everything to
the correct libs directory seems like the easiest solution.

Other file paths were merely mislabeled, I think, simply written wrong.
Because I never saw them appear at the paths that had been written in
the spec file at any point in the build process, but files with the same
name do appear in nearby directories.

A few times, it seemed to me that the spec file was being unnecessarily
cautious about which files to include, such as by using the %dir
directive to include only a directory itself, but then manually
including all the files within. So I tried to simplify that.

There's a Firefox compilation flag that turns on cairo-gtk3, but the
Firefox libs do not actually get a gtk3 directory. I'm honestly not yet
entirely sure what's up with that.