Rewrite, rearrange, reformat the man page.
I generally tried to more closely follow the man page guidelines in
`man-pages(7)`, for example switching the date of the last revision to
ISO format, and using the `.PP` macro to separate paragraphs. I didn't
see anything recommending this, but it made sense to me for the commas
in synonymous options to not be bolded, since you don't type the comma.
In a blog, I did see a recommendation against having any blank lines in
the source file, but adding more vertical whitespace makes the source
file easier to read, and doesn't seem to stop `man` from rendering the
page correctly.

I thought I should probably use non-breaking hyphens for the command
name.

Some command line options were removed or the information was out-of-
date, as I noted in a bug report for the Fedora package of Firefox
stable:
<https://bugzilla.redhat.com/show_bug.cgi?id=1334025>

I tried to sort the rest of the command line options by how often
they're likely to be used, as best as I could guess anyway, even though
it would probably be easier to just use the GUI menus and buttons to do
most of this stuff... So, it seemed appropriate to separate general
usage options from the options more for web developers. Some options,
when used incorrectly, will default to instead opening the home page in
a new window, so I tried to test for and explain the most likely errors
and Firefox's fall-back behavior.

Other sections, like "USAGE" and "BUGS" were sorta redundant; the
information is covered elsewhere.