Recommend using mock instead of rpmbuild in a VM.
Mock sandboxes the build process in a chroot, which means you can have a
clean separation of the build packages from the rest of your system
without necessarily needing to install a virtual machine. Building
outside of a VM also means compiling can go a bit faster.

The bit about running `git pull` on the spec repo seemed a little
pedantic in hindsight, and the `%tarball` directory macro is better
explained in the spec file comments, so I just removed those sections.